requirements.txt
Appium-Python-Client==3.1.0
setuptools
selenium==4.15.2
pylint==3.0.2